feat!: return pwd() or __source__.file dir if we reach file system root
I cannot think of any situation in which returning file system root
would be useful, so it makes more sense to return the
directory of the file called. This does depart a little
from what the here package does, which is to return
the working directory from which the script is called on the
command line if there is no match, but I think it makes more
sense to return the directory of the file, since the point is to
construct a path to root relative to the file in which
`@projectroot()` is called.